pyproject.toml
[project]
name = "discord-leveling-bot"
version = "1.0.0"
description = "Discord bot with XP leveling system and rank cards"
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord.py>=2.0.0",
    "aiosqlite>=0.17.0",
    "flask>=2.0.0",
    "pillow>=9.0.0",
    "aiohttp>=3.8.0"
]
